Move supported srs from wms to all layers
